Generalize the kernel function in PSetSoa
